+++ backend/storage/src/models/torrent_report.rs
@@ -9,12 +9,12 @@ pub struct TorrentReport {
     #[schema(value_type = String, format = DateTime)]
     pub reported_at: DateTime<Local>,
     pub reported_by_id: i32,
-    pub reported_torrent_id: i64,
+    pub reported_torrent_id: i32,
     pub description: String,
 }
 
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct UserCreatedTorrentReport {
-    pub reported_torrent_id: i64,
+    pub reported_torrent_id: i32,
     pub description: String,
 }
