+++ backend/storage/src/models/torrent_request.rs
@@ -20,14 +20,14 @@ use super::torrent::{
 #[derive(Debug, Serialize, Deserialize, FromRow, ToSchema)]
 pub struct TorrentRequest {
     pub id: i64,
-    pub title_group_id: i64,
+    pub title_group_id: i32,
     #[schema(value_type = String, format = DateTime)]
     pub created_at: DateTime<Utc>,
     #[schema(value_type = String, format = DateTime)]
     pub updated_at: DateTime<Utc>,
     pub created_by_id: i32,
     pub filled_by_user_id: Option<i32>,
-    pub filled_by_torrent_id: Option<i64>,
+    pub filled_by_torrent_id: Option<i32>,
     #[schema(value_type = String, format = DateTime)]
     pub filled_at: Option<DateTime<Utc>>,
     pub edition_name: Option<String>,
@@ -52,7 +52,7 @@ pub struct TorrentRequest {
 
 #[derive(Debug, Serialize, Deserialize, FromRow, ToSchema)]
 pub struct UserCreatedTorrentRequest {
-    pub title_group_id: i64,
+    pub title_group_id: i32,
     pub edition_name: Option<String>,
     pub release_group: Option<String>,
     pub description: Option<String>,
@@ -73,7 +73,7 @@ pub struct UserCreatedTorrentRequest {
 #[derive(Debug, Serialize, Deserialize, FromRow, ToSchema)]
 pub struct EditedTorrentRequest {
     pub id: i64,
-    pub title_group_id: i64,
+    pub title_group_id: i32,
     pub edition_name: Option<String>,
     pub release_group: Option<String>,
     pub description: Option<String>,
@@ -116,7 +116,7 @@ pub struct TorrentRequestWithTitleGroupLite {
 #[derive(Debug, Serialize, Deserialize, FromRow, ToSchema)]
 pub struct TorrentRequestFill {
     pub torrent_request_id: i64,
-    pub torrent_id: i64,
+    pub torrent_id: i32,
 }
 
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
